fix bitshift sign for pre-1.12 go support

+++ ft232h_pin_test.go
@@ -8,8 +8,8 @@ func TestPin(t *testing.T) {
 
 		ok := i > -1 && i < 8
 
-		d := D(i)
-		c := C(i)
+		d := D(uint(i))
+		c := C(uint(i))
 
 		if ok {
 
@@ -27,10 +27,10 @@ func TestPin(t *testing.T) {
 				t.Fatalf("C(%d) mask={%08b}, expected={%08b}", i, c.Mask(), 1<<i)
 			}
 
-			if i != d.Pos() {
+			if i != int(d.Pos()) {
 				t.Fatalf("D(%d) pos={%d}, expected={%d}", i, d.Pos(), i)
 			}
-			if i != c.Pos() {
+			if i != int(c.Pos()) {
 				t.Fatalf("C(%d) pos={%d}, expected={%d}", i, c.Pos(), i)
 			}
